P4ADEV-4297: moved PaymentsReportingLegacySoapServiceImplTest
